requirements.txt
# Enhanced requirements.txt - Complete Dependencies

# Core FastAPI and web dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# OpenAI for AI responses
openai==1.3.0

# Neo4j for graph database
neo4j==5.14.1

# Environment and configuration
python-dotenv==1.0.0

# Web scraping and HTTP requests
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Text processing and embeddings
sentence-transformers==2.2.2

# Async operations
asyncio-throttle==1.0.2

# Production server
gunicorn==21.2.0

# Optional: For better HTML parsing
html5lib==1.1